Replace old structure with new one
