+++ src/pedalo_server/__main__.py
@@ -1,0 +1,5 @@
+from .server import main
+
+if __name__ == "__main__":
+    print("Hello")
+    main()
