+++ src/pedalo_server/sensors/sensor.py
